+++ module/Batch/src/Command/MinerPayments/MinerPaymentsFactory.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Batch\Command\MinerPayments;
+
+class MinerPaymentsFactory {
+
+    public function __invoke($controllers): MinerPayments
+    {
+        return new MinerPayments($controllers->get('faucetdev'));
+    }
+}