+++ module/Batch/src/Command/MinerShares/MinerSharesFactory.php
@@ -4,7 +4,7 @@ namespace Batch\Command\MinerShares;
 
 class MinerSharesFactory
 {
-    public function __invoke($controllers)
+    public function __invoke($controllers): MinerShares
     {
         return new MinerShares($controllers->get('faucetdev'));
     }