+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes to this project will be documented in this file, in reverse chronological order by release.
 
+## Todo
+- Add "hacker" achievement to MinerShares
+
+## 1.1.0
+
+### Added
+- Miner Shares Batch
+- Transaction Tools
+
 ## 1.0.0
 
 ### Added
 - User Online Batch
+- Security Tools